chore: modify aditional file names following commit #7c230af46ddf9c
