+++ web-apps/client/src/components/MessageItem.jsx
@@ -1,56 +1,83 @@
-import React, {useContext} from 'react';
+import React from 'react';
+import {bool, string} from 'prop-types';
 
 // material ui
-import ListItem from '@material-ui/core/ListItem';
+import Paper from '@material-ui/core/Paper';
 import Grid from '@material-ui/core/Grid';
-
-// components
-import ChatBox from './ChatBox';
+import Avatar from '@material-ui/core/Avatar';
+import Typography from '@material-ui/core/Typography';
+import {makeStyles} from '@material-ui/core/styles';
 
 // utils
-import {getISOCurrentDate} from '../utils';
+import {getKoKRFormatDayOfWeek, getISOCurrentDate} from '../utils';
 
-// context
-import {UserContext} from '../contexts/User';
-
-// types
-import {messageShape} from '../types';
+// material ui style
+const useStyles = makeStyles((theme) => ({
+  root: {
+    flexGrow: 1,
+    overflow: 'hidden',
+    padding: theme.spacing(0, 3),
+  },
+  paper: {
+    maxWidth: 400,
+    margin: `${theme.spacing(1)}px`,
+    padding: theme.spacing(1),
+    fontWeight: 'bold',
+  },
+  colorPaper: {
+    backgroundColor: 'grey',
+    color: 'white',
+  },
+}));
 
 // component
-const MessageItem = ({message: {userId, content, timestamp}}) => {
-  // @ts-ignore
-  const {user} = useContext(UserContext);
+const MessageItem = ({content, isMyChat, timestamp}) => {
+  const classes = useStyles({});
 
   return (
-    <ListItem>
-      <Grid
-        container
-        direction='row'
-        justify={user === userId ? 'flex-end' : 'flex-start'}
-        alignItems='center'
-      >
+    <div className={classes.root}>
+      <Grid container alignItems='center'>
+        {!isMyChat && (
+          <Grid item>
+            <Avatar>상대방</Avatar>
+          </Grid>
+        )}
         <Grid item>
-          <ChatBox
-            content={content}
-            timestamp={timestamp}
-            isMyChat={user === userId}
-          />
+          <Grid
+            container
+            direction={!isMyChat ? 'row' : 'row-reverse'}
+            alignItems='flex-end'
+          >
+            <Grid item>
+              <Paper
+                className={`${classes.paper} ${!isMyChat &&
+                  classes.colorPaper}`}
+              >
+                <Typography>{content}</Typography>
+              </Paper>
+            </Grid>
+            <Grid item>
+              <Typography variant='caption'>
+                {getKoKRFormatDayOfWeek(timestamp)}
+              </Typography>
+            </Grid>
+          </Grid>
         </Grid>
       </Grid>
-    </ListItem>
+    </div>
   );
 };
 
 MessageItem.propTypes = {
-  message: messageShape,
+  content: string,
+  isMyChat: bool,
+  timestamp: string,
 };
 
 MessageItem.defaultProps = {
-  message: {
-    timestamp: getISOCurrentDate(),
-    content: '',
-    userId: null,
-  },
+  content: '',
+  isMyChat: false,
+  timestamp: getISOCurrentDate(),
 };
 
 export default MessageItem;
